refactor: BlackjackConfig 상수를 각 도메인 객체로 이동
